agents/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
aio-pika==9.3.1
asyncpg==0.29.0
redis==5.0.1
pydantic==2.5.0
python-dotenv==1.0.0
aiohttp==3.9.1
# OpenTelemetry packages
opentelemetry-api>=1.22.0
opentelemetry-sdk>=1.22.0
opentelemetry-instrumentation-aiohttp>=0.43b0
opentelemetry-instrumentation-asyncpg>=0.43b0
opentelemetry-instrumentation-aio-pika>=0.43b0
opentelemetry-exporter-otlp-proto-grpc>=1.22.0
opentelemetry-exporter-prometheus>=1.12.0
opentelemetry-semantic-conventions>=1.22.0
prometheus-client>=0.19.0
jsonschema>=4.20.0
pyyaml>=6.0
